feat: new generic form for remu csp and coef
